Cleaning FunctionSpace directory: better interface + mispelling correction of From into Form
git-svn-id: https://geuz.org/svn/gmsh/trunk@16271 59c8cabf-10d6-4a28-b43f-3936b85956c4

+++ projects/small_fem/formulation/TermCurlCurl.h
@@ -14,7 +14,7 @@
 
 class TermCurlCurl: public Term{
  private:
-  typedef const fullMatrix<double>& (Basis::*bFunction)(unsigned int s)const;
+  typedef const fullMatrix<double>& (Basis::*bFunction)(size_t s)const;
 
  public:
   TermCurlCurl(const GroupOfJacobian& goj,
