+++ projects/small_fem/formulation/TermGradGrad.h
@@ -14,7 +14,7 @@
 
 class TermGradGrad: public Term{
  private:
-  typedef const fullMatrix<double>& (Basis::*bFunction)(unsigned int s)const;
+  typedef const fullMatrix<double>& (Basis::*bFunction)(size_t s)const;
 
  public:
   TermGradGrad(const GroupOfJacobian& goj,
